third_party/blink/web_tests/external/wpt/service-workers/service-worker/xhr-content-length.https.window-expected.txt

This is a testharness.js-based test.
[FAIL] Synthetic response with two Content-Length headers value larger than response body length
  assert_equals: expected "10000, 10000" but got "10000"
Harness: the test ran to completion.
